--- edc-extensions/transferprocess-sftp-client/src/main/java/net/catenax/edc/transferprocess/sftp/client/SftpDataSink.java
@@ -1,29 +1,0 @@
-/*
- *  Copyright (c) 2022 Mercedes-Benz Tech Innovation GmbH
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Mercedes-Benz Tech Innovation GmbH - Initial Test
- *
- */
-
-package net.catenax.edc.transferprocess.sftp.client;
-
-import org.eclipse.dataspaceconnector.dataplane.spi.pipeline.DataSource;
-import org.eclipse.dataspaceconnector.dataplane.spi.pipeline.ParallelSink;
-import org.eclipse.dataspaceconnector.spi.response.StatusResult;
-
-import java.util.List;
-
-public class SftpDataSink extends ParallelSink {
-
-    @Override
-    protected StatusResult<Void> transferParts(List<DataSource.Part> parts) {
-        return null;
-    }
-}
