+++ edc-extensions/transferprocess-sftp-common/src/main/java/org/eclipse/tractusx/edc/trasnferprocess/sftp/common/SftpLocationFactory.java
@@ -12,7 +12,7 @@
  *
  */
 
-package net.catenax.edc.trasnferprocess.sftp.common;
+package org.eclipse.tractusx.edc.trasnferprocess.sftp.common;
 
 public interface SftpLocationFactory {
     SftpLocation createSftpLocation(String transferProcessId);
