+++ edc-extensions/transferprocess-sftp-common/src/main/java/org/eclipse/tractusx/edc/trasnferprocess/sftp/common/SftpUserFactory.java
@@ -12,21 +12,8 @@
  *
  */
 
-package net.catenax.edc.trasnferprocess.sftp.common;
+package org.eclipse.tractusx.edc.trasnferprocess.sftp.common;
 
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.ToString;
-
-import java.security.KeyPair;
-
-@Builder
-@Getter
-@ToString(of = "name")
-@EqualsAndHashCode
-public class SftpUser {
-    private final String name;
-    private final String password;
-    private final KeyPair keyPair;
+public interface SftpUserFactory {
+    SftpUser createSftpUser(String transferProcessId);
 }
