--- edc-extensions/transferprocess-sftp-provisioner/src/main/java/net/catenax/edc/transferprocess/sftp/provisioner/SftpProviderResourceDefinitionGenerator.java
@@ -1,34 +1,0 @@
-/*
- *  Copyright (c) 2022 Mercedes-Benz Tech Innovation GmbH
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Mercedes-Benz Tech Innovation GmbH - Initial API and Implementation
- *
- */
-
-package net.catenax.edc.transferprocess.sftp.provisioner;
-
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-import org.eclipse.dataspaceconnector.policy.model.Policy;
-import org.eclipse.dataspaceconnector.spi.transfer.provision.ProviderResourceDefinitionGenerator;
-import org.eclipse.dataspaceconnector.spi.types.domain.DataAddress;
-import org.eclipse.dataspaceconnector.spi.types.domain.transfer.DataRequest;
-import org.eclipse.dataspaceconnector.spi.types.domain.transfer.ResourceDefinition;
-import org.jetbrains.annotations.Nullable;
-
-@RequiredArgsConstructor
-public class SftpProviderResourceDefinitionGenerator implements ProviderResourceDefinitionGenerator {
-    @NonNull
-    private final String dataAddressType;
-    @Override
-    public @Nullable ResourceDefinition generate(DataRequest dataRequest, DataAddress assetAddress, Policy policy) {
-        return null;
-    }
-}
