+++ edc-extensions/transferprocess-sftp-provisioner/src/main/java/org/eclipse/tractusx/edc/transferprocess/sftp/provisioner/ConfigBackedSftpLocationFactory.java
@@ -12,12 +12,12 @@
  *
  */
 
-package net.catenax.edc.transferprocess.sftp.provisioner;
+package org.eclipse.tractusx.edc.transferprocess.sftp.provisioner;
 
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
-import net.catenax.edc.trasnferprocess.sftp.common.SftpLocation;
-import net.catenax.edc.trasnferprocess.sftp.common.SftpLocationFactory;
+import org.eclipse.tractusx.edc.trasnferprocess.sftp.common.SftpLocation;
+import org.eclipse.tractusx.edc.trasnferprocess.sftp.common.SftpLocationFactory;
 
 @RequiredArgsConstructor
 public class ConfigBackedSftpLocationFactory implements SftpLocationFactory {
