+++ edc-extensions/transferprocess-sftp-provisioner/src/main/java/org/eclipse/tractusx/edc/transferprocess/sftp/provisioner/NoopSftpProvider.java
@@ -12,11 +12,11 @@
  *
  */
 
-package net.catenax.edc.transferprocess.sftp.provisioner;
+package org.eclipse.tractusx.edc.transferprocess.sftp.provisioner;
 
-import net.catenax.edc.trasnferprocess.sftp.common.SftpLocation;
-import net.catenax.edc.trasnferprocess.sftp.common.SftpProvider;
-import net.catenax.edc.trasnferprocess.sftp.common.SftpUser;
+import org.eclipse.tractusx.edc.trasnferprocess.sftp.common.SftpLocation;
+import org.eclipse.tractusx.edc.trasnferprocess.sftp.common.SftpProvider;
+import org.eclipse.tractusx.edc.trasnferprocess.sftp.common.SftpUser;
 
 public class NoopSftpProvider implements SftpProvider {
     @Override
