+++ edc-extensions/transferprocess-sftp-provisioner/src/main/java/org/eclipse/tractusx/edc/transferprocess/sftp/provisioner/SftpProvisioner.java
@@ -12,15 +12,15 @@
  *
  */
 
-package net.catenax.edc.transferprocess.sftp.provisioner;
+package org.eclipse.tractusx.edc.transferprocess.sftp.provisioner;
 
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
-import net.catenax.edc.trasnferprocess.sftp.common.SftpLocation;
-import net.catenax.edc.trasnferprocess.sftp.common.SftpLocationFactory;
-import net.catenax.edc.trasnferprocess.sftp.common.SftpProvider;
-import net.catenax.edc.trasnferprocess.sftp.common.SftpUser;
-import net.catenax.edc.trasnferprocess.sftp.common.SftpUserFactory;
+import org.eclipse.tractusx.edc.trasnferprocess.sftp.common.SftpLocation;
+import org.eclipse.tractusx.edc.trasnferprocess.sftp.common.SftpLocationFactory;
+import org.eclipse.tractusx.edc.trasnferprocess.sftp.common.SftpProvider;
+import org.eclipse.tractusx.edc.trasnferprocess.sftp.common.SftpUser;
+import org.eclipse.tractusx.edc.trasnferprocess.sftp.common.SftpUserFactory;
 import org.eclipse.dataspaceconnector.policy.model.Policy;
 import org.eclipse.dataspaceconnector.spi.policy.engine.PolicyEngine;
 import org.eclipse.dataspaceconnector.spi.response.ResponseStatus;
